server/rag/requirements3.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.0
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
async-timeout==4.0.3
attrs==24.2.0
beautifulsoup4==4.12.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
dataclasses-json==0.6.7
exceptiongroup==1.2.2
frozenlist==1.5.0
google-ai-generativelanguage==0.6.10
google-api-core==2.23.0
google-api-python-client==2.153.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.66.0
grpcio==1.67.1
grpcio-status==1.67.1
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.0
idna==3.10
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.7
langchain-community==0.3.5
langchain-core==0.3.18
langchain-google-genai==2.0.4
langchain-text-splitters==0.3.2
langsmith==0.1.143
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
orjson==3.10.11
packaging==24.2
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyparsing==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
setuptools-rust==1.10.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.35
tenacity==9.0.0
threadpoolctl==3.5.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
yarl==1.17.1
